misc(ebpf): move get_path_str to shared header
